Extract projectLocation field
