+++ controllers/apply/awards-for-all/fields.js
@@ -1,9 +1,7 @@
 'use strict';
-const config = require('config');
-const moment = require('moment');
 const flatMap = require('lodash/flatMap');
 const get = require('lodash/fp/get');
-const has = require('lodash/has');
+const moment = require('moment');
 const { oneLine } = require('common-tags');
 
 const Joi = require('../lib/joi-extensions');
@@ -25,8 +23,6 @@ const {
     FREE_TEXT_MAXLENGTH
 } = require('./constants');
 
-const locationOptions = require('../lib/location-options');
-
 const fieldContactLanguagePreference = require('./fields/contact-language-preference');
 const fieldOrganisationStartDate = require('./fields/organisation-start-date');
 const fieldOrganisationType = require('./fields/organisation-type');
@@ -37,6 +33,7 @@ const fieldTotalIncomeYear = require('./fields/total-income-year');
 const fieldYourIdeaCommunity = require('./fields/your-idea-community');
 const fieldYourIdeaPriorities = require('./fields/your-idea-priorities');
 const fieldYourIdeaProject = require('./fields/your-idea-project');
+const fieldProjectLocation = require('./fields/project-location');
 
 module.exports = function fieldsFor({ locale, data = {} }) {
     const localise = get(locale);
@@ -753,63 +750,7 @@ module.exports = function fieldsFor({ locale, data = {} }) {
         },
         projectDateRange: fieldProjectDateRange(),
         projectCountry: fieldProjectCountry(locale),
-        projectLocation: {
-            name: 'projectLocation',
-            label: localise({
-                en: 'Where will your project take place?',
-                cy: 'Lle bydd eich prosiect wedi’i leoli? '
-            }),
-            explanation: localise({
-                en: oneLine`If your project covers more than one area please
-                    tell us where most of it will take place`,
-                cy: oneLine`Os yw eich prosiect mewn mwy nag un ardal, dywedwch
-                    wrthym lle bydd y rhan fwyaf ohono yn cymryd lle.`
-            }),
-            type: 'select',
-            defaultOption: localise({
-                en: 'Select a location',
-                cy: 'Dewiswch leoliad'
-            }),
-            get optgroups() {
-                const country = get('projectCountry')(data);
-                const locations = locationOptions(locale);
-
-                let result;
-                if (country === 'england') {
-                    result = locations.england;
-                } else if (country === 'scotland') {
-                    result = locations.scotland;
-                } else if (country === 'northern-ireland') {
-                    result = locations.northernIreland;
-                } else if (country === 'wales') {
-                    result = locations.wales;
-                } else {
-                    result = [];
-                }
-
-                return result;
-            },
-            isRequired: true,
-            get schema() {
-                const options = flatMap(this.optgroups, group => group.options);
-                return Joi.when('projectCountry', {
-                    is: Joi.exist(),
-                    then: Joi.string()
-                        .valid(options.map(option => option.value))
-                        .required(),
-                    otherwise: Joi.any().strip()
-                });
-            },
-            messages: [
-                {
-                    type: 'base',
-                    message: localise({
-                        en: 'Select a location',
-                        cy: 'Dewiswch leoliad'
-                    })
-                }
-            ]
-        },
+        projectLocation: fieldProjectLocation(locale, data),
         projectLocationDescription: {
             name: 'projectLocationDescription',
             label: localise({
